Remove hsTargets.rda and documentation, as that was only for the BioC08 tutorial. Start converting the start/end field to integer, rather than numeric. Fix bug in URL encoding of GFF output. Fix range calculation in scaling of genomeSegment objects. Improve efficiency of BED parsing when there is no track line. Add exercises to vignette.
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/rtracklayer@33107 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ R/trackSet.R
@@ -154,8 +154,8 @@ setMethod("trackFeatureData", "data.frame",
             if (is.null(object$strand))
               object$strand <- NA
             object$chrom <- paste("chr", sub("chr", "", object$chrom), sep = "")
-            object$start <- as.numeric(as.character(object$start))
-            object$end <- as.numeric(as.character(object$end))
+            object$start <- as.integer(as.character(object$start))
+            object$end <- as.integer(as.character(object$end))
             features <- as(object, "AnnotatedDataFrame")
             descs <- c(chrom = "Chromosome ID",
                        start = "Start position on chromosome",
